requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
python-dotenv>=0.14.0
mysqlclient>=2.0.1
mysql-connector-python>=8.0.21
